requirements.txt
numpy==2.2.0
pandas==2.2.3
Pillow==11.0.0
streamlit==1.41.1
tensorflow==2.15.0
tensorflow_macos==2.15.0
